+++ cmd/catalog.go
@@ -5,9 +5,6 @@ import (
 	"log"
 	"strings"
 
-	"github.com/docker/libcompose/docker"
-	"github.com/docker/libcompose/docker/ctx"
-	"github.com/docker/libcompose/project"
 	"github.com/spf13/cobra"
 )
 
@@ -37,22 +34,11 @@ func init() {
 // catalog images
 func catalog(cmd *cobra.Command, args []string) {
 
-	//read the harbor compose file
-	harborCompose := DeserializeHarborCompose(HarborComposeFile)
-
-	//use libcompose to parse yml file
-	dockerComposeProject, err := docker.NewProject(&ctx.Context{
-		Context: project.Context{
-			ComposeFiles: []string{DockerComposeFile},
-		},
-	}, nil)
-
-	if err != nil {
-		log.Fatal("error parsing compose file" + err.Error())
-	}
+	//read the compose files
+	dockerCompose, harborCompose := unmarshalComposeFiles(DockerComposeFile, HarborComposeFile)
 
 	//validate the compose file
-	_, err = dockerComposeProject.Config()
+	_, err := dockerCompose.Config()
 	if err != nil {
 		log.Fatal("error parsing compose file" + err.Error())
 	}
@@ -65,7 +51,7 @@ func catalog(cmd *cobra.Command, args []string) {
 		for _, containerName := range shipment.Containers {
 
 			//lookup the container in the list of services in the docker-compose file
-			serviceConfig, found := dockerComposeProject.GetServiceConfig(containerName)
+			serviceConfig, found := dockerCompose.GetServiceConfig(containerName)
 			if !found {
 				log.Fatal("could not find service in docker compose file")
 			}