+++ cmd/compose.go
@@ -1,33 +1,29 @@
 package cmd
 
 import (
+	"fmt"
 	"log"
+	"os"
 
 	yaml "gopkg.in/yaml.v2"
 
+	"github.com/docker/libcompose/config"
 	"github.com/docker/libcompose/docker"
 	"github.com/docker/libcompose/docker/ctx"
 	"github.com/docker/libcompose/project"
 )
 
-//unmarshal docker compose yaml
-func unmarshalDockerCompose(yamlString string) (DockerCompose, project.APIProject) {
+//unmarshal docker compose yaml string into a compose APIProject
+func unmarshalDockerCompose(yamlString string) project.APIProject {
 	if Verbose {
 		log.Printf("unmarshalDockerCompose - %v", yamlString)
 	}
 
 	yamlBits := []byte(yamlString)
 
-	//parse the docker compose file (only used for writing used by generate)
-	var dockerCompose DockerCompose
-	err := yaml.Unmarshal(yamlBits, &dockerCompose)
-	if err != nil {
-		log.Fatalf("error: %v", err)
-	}
-
 	//use libcompose to parse compose yml
 	bytes := [][]byte{yamlBits}
-	dockerComposeProject, err := docker.NewProject(&ctx.Context{
+	dockerCompose, err := docker.NewProject(&ctx.Context{
 		Context: project.Context{
 			ComposeBytes: bytes,
 			ProjectName:  "required",
@@ -37,7 +33,7 @@ func unmarshalDockerCompose(yamlString string) (DockerCompose, project.APIProjec
 		log.Fatal(err)
 	}
 
-	return dockerCompose, dockerComposeProject
+	return dockerCompose
 }
 
 //unmarshal harbor compose yaml
@@ -51,9 +47,25 @@ func unmarshalHarborCompose(yamlString string) HarborCompose {
 	return harborCompose
 }
 
-//unmarshals both docker compose and harbor compose yaml
+//unmarshals both docker compose and harbor compose yaml strings
 func unmarshalCompose(dockerComposeYaml string, harborComposeYaml string) (project.APIProject, HarborCompose) {
-	_, dc := unmarshalDockerCompose(dockerComposeYaml)
+	dc := unmarshalDockerCompose(dockerComposeYaml)
 	hc := unmarshalHarborCompose(harborComposeYaml)
 	return dc, hc
+}
+
+//unmarshals both docker compose and harbor compose yaml files
+func unmarshalComposeFiles(dockerComposeFile string, harborComposeFile string) (project.APIProject, HarborCompose) {
+	dc := DeserializeDockerCompose(dockerComposeFile)
+	hc := DeserializeHarborCompose(harborComposeFile)
+	return dc, hc
+}
+
+func getDockerComposeService(dockerCompose project.APIProject, container string) *config.ServiceConfig {
+	serviceConfig, success := dockerCompose.GetServiceConfig(container)
+	if !success {
+		fmt.Printf("ERROR: Container: %v defined in %v cannot be found in %v\n", container, HarborComposeFile, DockerComposeFile)
+		os.Exit(-1)
+	}
+	return serviceConfig
 }