+++ cmd/generate.go
@@ -28,11 +28,14 @@ Examples:
 harbor-compose generate my-shipment dev --build-provider local
 harbor-compose generate my-shipment dev -b circleciv1
 harbor-compose generate my-shipment dev -b circleciv2
+harbor-compose generate my-shipment dev -b codeship
 `,
 	Run: generate,
 }
 
 var buildProvider string
+
+const providerEc2 = "ec2"
 
 func init() {
 	generateCmd.PersistentFlags().StringVarP(&buildProvider, "build-provider", "b", "", "generate build provider-specific files that allow you to build Docker images do CI/CD with Harbor")
@@ -97,9 +100,7 @@ func generate(cmd *cobra.Command, args []string) {
 	}
 
 	username, token, err := Login()
-	if err != nil {
-		log.Fatal(err)
-	}
+	check(err)
 
 	shipment := args[0]
 	env := args[1]
@@ -122,13 +123,10 @@ func generate(cmd *cobra.Command, args []string) {
 	//if build provider is specified, allow it modify the compose objects and do its thing
 	if len(buildProvider) > 0 {
 		provider, err := getBuildProvider(buildProvider)
-		if err != nil {
-			log.Fatal(err)
-		}
+		check(err)
+
 		artifacts, err := provider.ProvideArtifacts(&dockerCompose, &harborCompose, shipmentObject.BuildToken)
-		if err != nil {
-			log.Fatal(err)
-		}
+		check(err)
 
 		//write artifacts to file system
 		if artifacts != nil {
@@ -136,18 +134,19 @@ func generate(cmd *cobra.Command, args []string) {
 				//create directories if needed
 				dirs := filepath.Dir(artifact.FilePath)
 				err = os.MkdirAll(dirs, os.ModePerm)
-				if err != nil {
-					log.Fatal(err)
-				}
+				check(err)
+
 				if _, err := os.Stat(artifact.FilePath); err == nil {
 					//exists
 					fmt.Print(artifact.FilePath + " already exists. Overwrite? ")
 					if askForConfirmation() {
 						err = ioutil.WriteFile(artifact.FilePath, []byte(artifact.FileContents), 0644)
+						check(err)
 					}
 				} else {
 					//doesn't exist
 					err = ioutil.WriteFile(artifact.FilePath, []byte(artifact.FileContents), 0644)
+					check(err)
 				}
 			}
 		}
@@ -180,17 +179,6 @@ func generate(cmd *cobra.Command, args []string) {
 	}
 }
 
-//find the ec2 provider
-func ec2Provider(providers []ProviderPayload) *ProviderPayload {
-	for _, provider := range providers {
-		if provider.Name == "ec2" {
-			return &provider
-		}
-	}
-	log.Fatal("ec2 provider is missing")
-	return nil
-}
-
 func transformShipmentToHarborCompose(shipmentObject *ShipmentEnvironment, dockerCompose *DockerCompose) HarborCompose {
 
 	//convert a Shipment object into a HarborCompose object with a single shipment
@@ -213,7 +201,6 @@ func transformShipmentToHarborCompose(shipmentObject *ShipmentEnvironment, docke
 	//environment
 	copyEnvVars(shipmentObject.EnvVars, nil, special)
 
-	//look for the ec2 provider (for now)
 	provider := ec2Provider(shipmentObject.Providers)
 
 	//now populate other harbor-compose metadata