+++ cmd/ps.go
@@ -31,8 +31,7 @@ func init() {
 func ps(cmd *cobra.Command, args []string) {
 
 	//read the compose files
-	_, dockerCompose := DeserializeDockerCompose(DockerComposeFile)
-	harborCompose := DeserializeHarborCompose(HarborComposeFile)
+	dockerCompose, harborCompose := unmarshalComposeFiles(DockerComposeFile, HarborComposeFile)
 
 	doPs(dockerCompose, harborCompose)
 }
@@ -154,7 +153,7 @@ func getShipmentPrimaryPort(dockerCompose project.APIProject, shipment ComposeSh
 	for _, container := range shipment.Containers {
 		serviceConfig, success := dockerCompose.GetServiceConfig(container)
 		if !success {
-			return "", errors.New("error getting service config")
+			return "", errors.New("unabled to find container in docker-compose.yml")
 		}
 
 		if serviceConfig.Ports == nil {